refactor: modularize buffer module into more separate files for easier reading and implementation
